Update reservation part
Finished everything
