+++ mobile/lib/presentation/screens/room_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hotel_flutter/presentation/widgets/room/background_room_image.dart'; // Ensure this file exists
-import 'package:hotel_flutter/presentation/widgets/room/room_details.dart'; // Ensure this file exists
+import 'package:hotel_flutter/presentation/widgets/room/reservation.dart'; // Ensure this file exists
 
 class RoomScreen extends StatefulWidget {
   final String backgroundImage;
@@ -33,7 +33,7 @@ class _RoomScreenState extends State<RoomScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Room Details'),
+        title: Text(widget.roomName),
         backgroundColor: Colors.transparent,
         elevation: 0,
         actions: [
@@ -55,7 +55,7 @@ class _RoomScreenState extends State<RoomScreen> {
               width: double.infinity,
               child: BackgroundRoomImage(image: widget.backgroundImage),
             ),
-            RoomDetails(
+            Reservation(
               roomName: widget.roomName,
               rating: widget.rating,
               price: widget.price,
@@ -68,7 +68,6 @@ class _RoomScreenState extends State<RoomScreen> {
                 });
               },
             ),
-            // Other buttons or widgets...
           ],
         ),
       ),
