+++ mobile/lib/presentation/widgets/room/day_selector.dart
@@ -3,12 +3,14 @@ import 'package:flutter/material.dart';
 class DaySelector extends StatelessWidget {
   final int daysInMonth;
   final int activeDay;
+  final ValueChanged<int> onDayTap;
 
   const DaySelector({
-    Key? key,
+    super.key,
     required this.daysInMonth,
     required this.activeDay,
-  }) : super(key: key);
+    required this.onDayTap,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -17,27 +19,32 @@ class DaySelector extends StatelessWidget {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.start,
         children: List.generate(daysInMonth, (index) {
-          return Container(
-            margin: const EdgeInsets.symmetric(horizontal: 4.0),
-            decoration: BoxDecoration(
-              color: activeDay == index + 1
-                  ? const Color.fromARGB(255, 29, 53, 115)
-                  : Colors.white,
-              borderRadius: BorderRadius.circular(10),
-              border: Border.all(
-                color: const Color.fromARGB(255, 142, 142, 147),
-                width: 1,
-              ),
-            ),
-            width: 40,
-            height: 40,
-            alignment: Alignment.center,
-            child: Text(
-              '${index + 1}',
-              style: TextStyle(
+          return GestureDetector(
+            onTap: () {
+              onDayTap(index + 1);
+            },
+            child: Container(
+              margin: const EdgeInsets.symmetric(horizontal: 4.0),
+              decoration: BoxDecoration(
                 color: activeDay == index + 1
-                    ? Colors.white
-                    : const Color.fromARGB(255, 29, 53, 115),
+                    ? const Color.fromARGB(255, 29, 53, 115)
+                    : const Color.fromARGB(255, 217, 217, 217),
+                borderRadius: BorderRadius.circular(10),
+                border: Border.all(
+                  color: const Color.fromARGB(255, 217, 217, 217),
+                  width: 1,
+                ),
+              ),
+              width: 40,
+              height: 40,
+              alignment: Alignment.center,
+              child: Text(
+                '${index + 1}',
+                style: TextStyle(
+                  color: activeDay == index + 1
+                      ? Colors.white
+                      : const Color.fromARGB(255, 127, 125, 125),
+                ),
               ),
             ),
           );
